+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ CHANGELOG
 
 Release NEXT
 ------------
+- Implement management command to cleanup invalid price estimates.
+- Update metadata for price estimates of service, settings and project on scope deletion.
 - Allow to update push hook token.
 
 Release 0.102.5